+++ tests/Integration/Service.Tests/AuthModuleSpecs/AccessTokenContext.cs
@@ -5,23 +5,34 @@
     using Linn.Common.Facade;
     using Linn.Tax.Domain;
     using Linn.Tax.Facade.ResourceBuilders;
+    using Linn.Tax.Proxy;
     using Linn.Tax.Resources;
     using Linn.Tax.Service.Modules;
+    using Linn.Tax.Service.ResponseProcessors;
 
     using Nancy;
     using Nancy.Session;
+    using Nancy.Testing;
+
+    using NSubstitute;
 
     using NUnit.Framework;
 
-    public class AccessTokenContext : ContextBase
+    public class AccessTokenContext : NancyContextBase
     {
+        protected IHmrcApiService ApiService { get; set; }
+
+        protected string RedirectLocation { get; set; }
+
         [SetUp]
-        public new void EstablishContext()
+        public void EstablishContext()
         {
-            this.Cfg = with =>
+            this.ApiService = Substitute.For<IHmrcApiService>();
+            var bootstrapper = new ConfigurableBootstrapper(with =>
                 {
                     with.Dependency(this.ApiService);
                     with.Dependency<IResourceBuilder<VatReturn>>(new VatReturnResourceBuilder());
+                    with.ResponseProcessor<VatReturnResponseProcessor>();
                     with.Module<AuthenticationModule>();
                     with.RequestStartup(
                         (container, pipelines, context) =>
@@ -37,8 +48,8 @@
                                         });
                                 pipelines.BeforeRequest = b;
                             });
-                };
-            base.EstablishContext();
+                });
+            this.Browser = new Browser(bootstrapper);
         }
     }
 }
