+++ tests/Integration/Service.Tests/AuthModuleSpecs/NoAccessTokenContext.cs
@@ -5,26 +5,35 @@
     using Linn.Common.Facade;
     using Linn.Tax.Domain;
     using Linn.Tax.Facade.ResourceBuilders;
+    using Linn.Tax.Proxy;
     using Linn.Tax.Resources;
     using Linn.Tax.Service.Modules;
+    using Linn.Tax.Service.ResponseProcessors;
 
     using Nancy;
     using Nancy.Session;
+    using Nancy.Testing;
 
     using NSubstitute;
 
     using NUnit.Framework;
 
-    public class NoAccessTokenContext : ContextBase
+    public class NoAccessTokenContext : NancyContextBase 
     {
+        protected IHmrcApiService ApiService { get; set; }
+
+        protected string RedirectLocation { get; set; }
+
         [SetUp]
-        public new void EstablishContext()
+        public void EstablishContext()
         {
+            this.ApiService = Substitute.For<IHmrcApiService>();
             this.ApiService.ExchangeCodeForAccessToken(Arg.Any<string>()).Returns(new TokenResource { access_token = "blah" });
-            this.Cfg = with =>
+            var bootstrapper = new ConfigurableBootstrapper(with =>
                 {
                     with.Dependency(this.ApiService);
                     with.Dependency<IResourceBuilder<VatReturn>>(new VatReturnResourceBuilder());
+                    with.ResponseProcessor<VatReturnResponseProcessor>();
                     with.Module<AuthenticationModule>();
                     with.RequestStartup(
                         (container, pipelines, context) =>
@@ -39,9 +48,9 @@
                                 pipelines.BeforeRequest = b;
                             });
                     with.ApplicationStartup((container, pipelines) => { });
-                };
+                });
 
-            base.EstablishContext();
+            this.Browser = new Browser(bootstrapper);
         }
     }
 }
