+++ tests/Integration/Service.Tests/VatReturnModuleSpecs/ContextBase.cs
@@ -1,11 +1,12 @@
 ﻿namespace Linn.Tax.Service.Tests.VatReturnModuleSpecs
 {
-    using System.Net;
-
-    using Linn.Tax.Facade;
+    using Linn.Common.Facade;
+    using Linn.Tax.Domain;
+    using Linn.Tax.Facade.ResourceBuilders;
     using Linn.Tax.Facade.Services;
     using Linn.Tax.Proxy;
     using Linn.Tax.Service.Modules;
+    using Linn.Tax.Service.ResponseProcessors;
 
     using Nancy.Session;
     using Nancy.Testing;
@@ -28,6 +29,7 @@
             var bootstrapper = new ConfigurableBootstrapper(
                 with =>
                     {
+                        with.Dependency<IResourceBuilder<VatReturn>>(new VatReturnResourceBuilder());
                         with.Dependency(this.ApiService);
                         with.Dependency(this.VatReturnService);
                         with.Module<VatReturnModule>();
